quilt-bot: Trigger code review assignment using webhooks
Before, we polled GitHub to look for any PRs that needed assignment.
However, now that quilt-bot handles all the repositories in the quilt
organization, we would get rate-limited by GitHub with this naive
approach. This patch migrates us to GitHub's webhooks API to only check PR
assignments when PRs are created or reviewed.